# admin/modules/iotserver/node/src/api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: '0.0.1'
  title: Sangoma IoT Server API
# during dev, should point to your local machine
host: 192.168.1.115:3000
basePath: /v1
securityDefinitions:
  basicAuth:
    type: basic
schemes:
  # tip: remove http to make production-grade
  - http
  #- https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /zones:
    x-swagger-router-controller: iot_server
    parameters:
      - name: sessionid
        in: query
        description: swagger session id
        required: false
        type: string
        default: 'swagger_session'
      - name: token
        in: query
        description: swagger token
        required: false
        type: string
        default: 'iZh8grpZbjQZ2JIFyViUfOoVLcOT7XYqZdu1u34frvfNQrL8ifJAHgpgV'
    get:
      description: Returns list of available zones
      operationId: getZones
      security: []
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/ArrayOfZones'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    post:
      description: Create a new zone
      operationId: createZone
      security: []
      consumes:
        - application/json
        - application/octet-stream
      parameters:
        - name: data
          in: body
          schema:
            type: object
            properties:
              name:
                type: string
                description: The zone name
              location_uuid:
                type: string
                description: The uuid of the location the zone belongs to
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Zone'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    delete:
      description: delete zone
      operationId: deleteZone
      security: []
      parameters:
        - name: uuid
          in: query
          description: The uuid of the zone to delete
          required: true
          type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/ArrayOfZones'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  /zones/{uuid}:
    x-swagger-router-controller: iot_server
    parameters:
      - name: sessionid
        in: query
        description: swagger session id
        required: false
        type: string
        default: 'swagger_session'
      - name: token
        in: query
        description: swagger token
        required: false
        type: string
        default: 'iZh8grpZbjQZ2JIFyViUfOoVLcOT7XYqZdu1u34frvfNQrL8ifJAHgpgV'
      - name: uuid
        in: path
        description: The uuid of the zone
        required: true
        type: string
    get:
      description: get a specific zone
      operationId: getZone
      security: []
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Zone'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    post:
      description: update a zone
      operationId: updateZone
      security: []
      consumes:
        - application/json
        - application/octet-stream
      parameters:
        - name: data
          in: body
          schema:
            type: array
            description: list parameters to update
            items:
              type: object
              properties:
                name:
                  type: string
                  description: name of the parameter
                value:
                  type: [ string, object, number, integer, boolean, array ]
                  description: value of the parameter
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Zone'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  /zones/{uuid}/permissions:
    x-swagger-router-controller: iot_server
    parameters:
      - name: sessionid
        in: query
        description: swagger session id
        required: false
        type: string
        default: 'swagger_session'
      - name: token
        in: query
        description: swagger token
        required: false
        type: string
        default: 'iZh8grpZbjQZ2JIFyViUfOoVLcOT7XYqZdu1u34frvfNQrL8ifJAHgpgV'
      - name: uuid
        in: path
        description: The uuid of the zone to add a permission to
        required: true
        type: string
    post:
      description: add a permission to a zone
      operationId: zoneAddPermission
      security: []
      consumes:
        - application/json
        - application/octet-stream
      parameters:
        - name: data
          in: body
          schema:
            type: object
            properties:
              user_group_id:
                type: integer
                description: The user group member of this permission
              permission_type_id:
                type: integer
                enum: [1, 2, 3, 4]
                description: The id of the permission type. 1 => view_limitted, 2 => view,  3 => control, 4 => monitor, 
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Zone'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    delete:
      description: remove a permission from a zone
      operationId: zoneRemovePermission
      security: []
      consumes:
        - application/json
        - application/octet-stream
      parameters:
        - name: permission_id
          in: query
          description: The id of the permission to remove from this location
          required: true
          type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/ArrayOfZones'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  /zones/{uuid}/scenes:
    x-swagger-router-controller: iot_server
    parameters:
      - name: sessionid
        in: query
        description: swagger session id
        required: false
        type: string
        default: 'swagger_session'
      - name: token
        in: query
        description: swagger token
        required: false
        type: string
        default: 'iZh8grpZbjQZ2JIFyViUfOoVLcOT7XYqZdu1u34frvfNQrL8ifJAHgpgV'
      - name: uuid
        in: path
        description: The uuid of the zone to remove the scenefrom
        required: true
        type: string
    post:
      description: add a scene to a zone
      operationId: zoneAddScene
      security: []
      consumes:
        - application/json
        - application/octet-stream
      parameters:
        - name: scene_uuid
          in: query
          description: The uuid of the scene to add to this location
          required: true
          type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Scene'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    delete:
      description: remove a scene from a zone
      operationId: zoneRemoveScene
      security: []
      consumes:
        - application/json
        - application/octet-stream
      parameters:
        - name: scene_uuid
          in: query
          description: The uuid of the scene to remove from this zone
          required: true
          type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/ArrayOfScenes'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  /zones/{uuid}/action:
    x-swagger-router-controller: iot_server
    parameters:
      - name: sessionid
        in: query
        description: swagger session id
        required: false
        type: string
        default: 'swagger_session'
      - name: token
        in: query
        description: swagger token
        required: false
        type: string
        default: 'iZh8grpZbjQZ2JIFyViUfOoVLcOT7XYqZdu1u34frvfNQrL8ifJAHgpgV'
      - name: uuid
        in: path
        description: The uuid of the zone
        required: true
        type: string
    post:
      description: perform a zone action
      operationId: zoneDoAction
      security: []
      consumes:
        - application/json
        - application/octet-stream
      parameters:
        - name: action
          in: query
          description: The action to perform
          required: true
          type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Response'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
...
definitions:
...
  Response:
     required:
       - message
     properties:
       message:
         type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  Zone:
    type: object
    properties: 
      uuid:
        type: string
        description: zone unique identifier
      name:
        type: string
        description: zone name
      actions:
        type: string
        description: zone allowed actions
      temperature:
        type: string
        description: zone temperature
      children:
        type: array
        description: list of scenes belonging to this zone
        items:
          type: object
          properties:
            name:
              type: string
              description: scene name
            uuid:
              type: string
              description: scene uuid
            actions:
              type: string
              description: scene allowed actions
            objtype:
              type: string
              description: scene object type
      objtype:
        type: string
        description: the object type
  ArrayOfZones:
    type: array
    items:
      type: object
      properties: 
        uuid:
          type: string
          description: zone unique identifier
        name:
          type: string
          description: zone name
        actions:
          type: string
          description: zone allowed actions
        temperature:
          type: string
          description: zone temperature
        objtype:
          type: string
          description: the object type
  Scene:
    type: object
    properties: 
      uuid:
        type: string
        description: scene unique identifier
      name:
        type: string
        description: scene name
      actions:
        type: string
        description: scene allowed actions
      temperature:
        type: string
        description: scene temperature
      children:
        type: array
        description: list of groups belonging to this scene
        items:
          type: object
          properties:
            name:
              type: string
              description: group name
            uuid:
              type: string
              description: group uuid
            actions:
              type: string
              description: group allowed actions
            objtype:
              type: string
              description: group object type
      objtype:
        type: string
        description: the object type
  ArrayOfScenes:
    type: array
    items:
      type: object
      properties: 
        uuid:
          type: string
          description: scene unique identifier
        name:
          type: string
          description: scene name
        actions:
          type: string
          description: scene allowed actions
        temperature:
          type: string
          description: scene temperature
        objtype:
          type: string
          description: the object type
...